fix(sessions): label the outcome on the badge the replay view actually renders
The enum mapping landed on helpers nothing calls, so the pill in the
Replay view still read "200 success". Move outcomeLabel (and the
humaniseAction it falls back to) into utils/outcome-label so a component
can say what happened without importing the activity feed, and use it in
session-replay-panel's per-message metrics, which is the path render()
reaches. activity-feed re-exports both names, so its callers and tests
are unchanged.

The new test mounts the panel in replay mode and reads the pill from the
rendered shadow root, so it fails if the badge stops being on screen.
